+++ meta_package_manager/xbar/__init__.py
@@ -15,7 +15,7 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
-""" This is the last place in which we support Python 2.7 to allow BitBar
+""" This is the last place in which we support Python 2.7 to allow xbar
 plugin to be tested out in our CI workflow.
 """
 
@@ -33,13 +33,13 @@ except ImportError:
     PY = 2
 
 
-# Manually import BitBar plugin source code as a module, because of its non
+# Manually import xbar plugin source code as a module, because of its non
 # Python-compliant filename with a double extension made the first dot
 # interpreted as a submodule.
 if PY == 2:
     here = path.dirname(path.abspath(__file__))
-    bitbar_plugin = path.join(here, "meta_package_manager.7h.py")
-    imp.load_source(__name__, bitbar_plugin)
+    xbar_plugin = path.join(here, "meta_package_manager.7h.py")
+    imp.load_source(__name__, xbar_plugin)
 else:
     SourceFileLoader(
         __name__,